add: link to discord server at header
