+++ src/components/home/Term.tsx
@@ -13,7 +13,7 @@ const Term: React.FC = () => {
           本アプリの開発者はユーザーが本アプリを使用したことにより生じる損害について、いかなる責任も負いません。
         </ListItem>
         <ListItem>
-          お問い合わせは
+          お問い合わせは開発者の
           <Link
             href="https://discord.gg/rct5sx6rbZ"
             isExternal
